backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
Pillow
gunicorn
dj-rest-auth
django-allauth
requests
cryptography 
google-auth
